chore(avatar-cropper): split demo
